Use NSUInteger instead of unsigned int and CGFloat instead of float
Needed to switch to using NSNotFound instead of -1 in internal _request?(Row|Col) ivars

+++ Pod/Classes/XKPhotoScrollView.h
@@ -23,19 +23,19 @@ typedef NS_ENUM(NSInteger, XKPhotoScrollViewAnimationType) {
 @interface XKPhotoScrollViewViewState : NSObject <NSCopying>
 
 @property (nonatomic, strong) UIView *view;
-@property (nonatomic, assign) float scale;
-@property (nonatomic, assign) float baseScale;
-@property (nonatomic, assign) unsigned int row, col;
+@property (nonatomic, assign) CGFloat scale;
+@property (nonatomic, assign) CGFloat baseScale;
+@property (nonatomic, assign) NSUInteger row, col;
 @property (nonatomic, assign) BOOL placeholder;
 
 @end
 
 @protocol XKPhotoScrollViewDataSource
 
-- (void)photoScrollView:(XKPhotoScrollView *)photoScrollView requestViewAtRow:(unsigned int)row col:(unsigned int)col;
-- (void)photoScrollView:(XKPhotoScrollView *)photoScrollView cancelRequestAtRow:(unsigned int)row col:(unsigned int)col;
-- (unsigned int)photoScrollViewRows:(XKPhotoScrollView *)photoScrollView;
-- (unsigned int)photoScrollViewCols:(XKPhotoScrollView *)photoScrollView;
+- (void)photoScrollView:(XKPhotoScrollView *)photoScrollView requestViewAtRow:(NSUInteger)row col:(NSUInteger)col;
+- (void)photoScrollView:(XKPhotoScrollView *)photoScrollView cancelRequestAtRow:(NSUInteger)row col:(NSUInteger)col;
+- (NSUInteger)photoScrollViewRows:(XKPhotoScrollView *)photoScrollView;
+- (NSUInteger)photoScrollViewCols:(XKPhotoScrollView *)photoScrollView;
 
 @end
 
@@ -43,15 +43,15 @@ typedef NS_ENUM(NSInteger, XKPhotoScrollViewAnimationType) {
 
 @optional
 
-- (void)photoScrollView:(XKPhotoScrollView *)photoScrollView didTapView:(UIView *)view atPoint:(CGPoint)pt atRow:(unsigned int)row atCol:(unsigned int)col;
-- (void)photoScrollView:(XKPhotoScrollView *)photoScrollView didLongPressView:(UIView *)view atPoint:(CGPoint)pt atRow:(unsigned int)row atCol:(unsigned int)col;
-- (void)photoScrollView:(XKPhotoScrollView *)photoScrollView didTouchView:(UIView *)view withTouches:(NSSet *)touches atRow:(unsigned int)row atCol:(unsigned int)col;
-- (void)photoScrollView:(XKPhotoScrollView *)photoScrollView didDragView:(UIView *)view atRow:(unsigned int)row atCol:(unsigned int)col;
+- (void)photoScrollView:(XKPhotoScrollView *)photoScrollView didTapView:(UIView *)view atPoint:(CGPoint)pt atRow:(NSUInteger)row atCol:(NSUInteger)col;
+- (void)photoScrollView:(XKPhotoScrollView *)photoScrollView didLongPressView:(UIView *)view atPoint:(CGPoint)pt atRow:(NSUInteger)row atCol:(NSUInteger)col;
+- (void)photoScrollView:(XKPhotoScrollView *)photoScrollView didTouchView:(UIView *)view withTouches:(NSSet *)touches atRow:(NSUInteger)row atCol:(NSUInteger)col;
+- (void)photoScrollView:(XKPhotoScrollView *)photoScrollView didDragView:(UIView *)view atRow:(NSUInteger)row atCol:(NSUInteger)col;
 - (void)photoScrollView:(XKPhotoScrollView *)photoScrollView didUpdateTransformationForView:(UIView *)view withState:(XKPhotoScrollViewViewState *)state;
-- (void)photoScrollView:(XKPhotoScrollView *)photoScrollView didZoomView:(UIView *)view atRow:(unsigned int)row atCol:(unsigned int)col;
-- (void)photoScrollView:(XKPhotoScrollView *)photoScrollView didPinchDismissView:(UIView *)view atRow:(unsigned int)row atCol:(unsigned int)col;
-- (void)photoScrollView:(XKPhotoScrollView *)photoScrollView didRotateTo:(float)rotation;
-- (void)photoScrollView:(XKPhotoScrollView *)photoScrollView didChangeToRow:(unsigned int)row col:(unsigned int)col;
+- (void)photoScrollView:(XKPhotoScrollView *)photoScrollView didZoomView:(UIView *)view atRow:(NSUInteger)row atCol:(NSUInteger)col;
+- (void)photoScrollView:(XKPhotoScrollView *)photoScrollView didPinchDismissView:(UIView *)view atRow:(NSUInteger)row atCol:(NSUInteger)col;
+- (void)photoScrollView:(XKPhotoScrollView *)photoScrollView didRotateTo:(CGFloat)rotation;
+- (void)photoScrollView:(XKPhotoScrollView *)photoScrollView didChangeToRow:(NSUInteger)row col:(NSUInteger)col;
 - (void)photoScrollView:(XKPhotoScrollView *)photoScrollView didSetCurrentView:(UIView *)view withState:(XKPhotoScrollViewViewState *)state;
 - (void)photoScrollView:(XKPhotoScrollView *)photoScrollView isStabilizing:(UIView *)view;
 - (void)photoScrollView:(XKPhotoScrollView *)photoScrollView didStabilizeView:(UIView *)view;
@@ -61,28 +61,28 @@ typedef NS_ENUM(NSInteger, XKPhotoScrollViewAnimationType) {
 @interface XKPhotoScrollView : UIView
 
 - (void)reloadData;
-- (void)setView:(UIView *)view atRow:(unsigned int)row col:(unsigned int)col placeholder:(BOOL)placeholder;
-- (void)setViewOnMainThread:(UIView *)view atRow:(unsigned int)row col:(unsigned int)col placeholder:(BOOL)placeholder;
-- (BOOL)wantsViewAtRow:(unsigned int)row col:(unsigned int)col;
-- (void)setViewScale:(float)scale;
-- (void)setCurrentRow:(unsigned int)row col:(unsigned int)col;
-- (void)setCurrentRow:(unsigned int)aRow col:(unsigned int)aCol animated:(BOOL)animated;
+- (void)setView:(UIView *)view atRow:(NSUInteger)row col:(NSUInteger)col placeholder:(BOOL)placeholder;
+- (void)setViewOnMainThread:(UIView *)view atRow:(NSUInteger)row col:(NSUInteger)col placeholder:(BOOL)placeholder;
+- (BOOL)wantsViewAtRow:(NSUInteger)row col:(NSUInteger)col;
+- (void)setViewScale:(CGFloat)scale;
+- (void)setCurrentRow:(NSUInteger)row col:(NSUInteger)col;
+- (void)setCurrentRow:(NSUInteger)aRow col:(NSUInteger)aCol animated:(BOOL)animated;
 - (void)notifyDeviceOrientationDidChange:(UIDeviceOrientation)orientation animated:(BOOL)animated;
 
 @property (weak, nonatomic) IBOutlet id<XKPhotoScrollViewDataSource> dataSource;
 @property (weak, nonatomic) IBOutlet id<XKPhotoScrollViewDelegate> delegate;
 @property (assign, nonatomic) BOOL bouncesZoom;
 @property (assign, nonatomic) BOOL alwaysBounceScroll;
-@property (assign, nonatomic) float maximumZoomScale;
-@property (assign, nonatomic) float minimumZoomScale;
-@property (assign, nonatomic) float maximumBaseScale;
-@property (assign, nonatomic) float minimumDrag;
-@property (readonly, nonatomic) unsigned int col;
-@property (readonly, nonatomic) unsigned int row;
+@property (assign, nonatomic) CGFloat maximumZoomScale;
+@property (assign, nonatomic) CGFloat minimumZoomScale;
+@property (assign, nonatomic) CGFloat maximumBaseScale;
+@property (assign, nonatomic) CGFloat minimumDrag;
+@property (readonly, nonatomic) NSUInteger col;
+@property (readonly, nonatomic) NSUInteger row;
 @property (readonly, nonatomic) int rotation;
 @property (readonly, nonatomic) BOOL touching;
-@property (assign, nonatomic) float viewScale;
-@property (readonly, nonatomic) float baseScale;
+@property (assign, nonatomic) CGFloat viewScale;
+@property (readonly, nonatomic) CGFloat baseScale;
 @property (assign, nonatomic) UIEdgeInsets insets;
 @property (assign, nonatomic) UIEdgeInsets baseInsets;
 @property (assign, nonatomic) CGPoint viewOffset;
